backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ruc-wenshu-backend"
version = "0.1.0"
description = "中国人民大学信息学院「文枢」大模型后端"
authors = [
    {name = "RUC Information School", email = "info@ruc.edu.cn"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    # LlamaIndex core
    "llama-index>=0.9.0",
    "llama-index-llms-google-genai>=0.1.0",
    "llama-index-llms-deepseek>=0.1.1",
    "llama-index-llms-google-genai>=0.2.1",
    "llama-index-llms-openai-like>=0.3.5",
    "llama-index-embeddings-openai-like>=0.1.0",
    # Docling support
    "llama-index-readers-docling>=0.1.0",
    "llama-index-readers-file>=0.4.8",
    "llama-index-node-parser-docling>=0.1.0",
    "docling>=2.36.1",
    "llama-index>=0.12.38",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "vllm>=0.9.0.1",
    "xlrd>=2.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.6.0",
]

[project.scripts]
wenshu = "wenshu.main:main"

[tool.hatch.build.targets.wheel]
packages = ["wenshu"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
